+++ src/app/auth/forms/signup-form/signup-form.component.ts
@@ -6,8 +6,8 @@ import { SelectOption } from 'app/shared/components/select-field/select-field.co
 import { UserRole } from 'app/core/models/user/user-role.model';
 import { SentenceCasePipe } from 'app/shared/pipes/sentence-case/sentence-case.pipe';
 import { SplitCamelCasePipe } from 'app/shared/pipes/split-camel-case/split-camel-case.pipe';
-import { EmailValidator } from 'app/shared/validators/email/email.validator';
-import { PasswordValidator } from 'app/shared/validators/password/password.validator';
+import { EmailValidator } from 'app/core/validators/email/email.validator';
+import { PasswordValidator } from 'app/core/validators/password/password.validator';
 
 @Component({
   selector: 'app-signup-form',
