+++ src/app/core/validators/email/email.validator.ts
